+++ include/r3d/r3d.h
@@ -1,0 +1,42 @@
+﻿/*
+ * Copyright (c) 2025 Le Juez Victor
+ *
+ * This software is provided "as-is", without any express or implied warranty. In no event
+ * will the authors be held liable for any damages arising from the use of this software.
+ *
+ * Permission is granted to anyone to use this software for any purpose, including commercial
+ * applications, and to alter it and redistribute it freely, subject to the following restrictions:
+ *
+ *   1. The origin of this software must not be misrepresented; you must not claim that you
+ *   wrote the original software. If you use this software in a product, an acknowledgment
+ *   in the product documentation would be appreciated but is not required.
+ *
+ *   2. Altered source versions must be plainly marked as such, and must not be misrepresented
+ *   as being the original software.
+ *
+ *   3. This notice may not be removed or altered from any source distribution.
+ */
+
+#ifndef R3D_H
+#define R3D_H
+
+#include "r3d_animation.h"
+#include "r3d_api.h"
+#include "r3d_core.h"
+#include "r3d_culling.h"
+#include "r3d_curves.h"
+#include "r3d_decal.h"
+#include "r3d_draw.h"
+#include "r3d_environment.h"
+#include "r3d_lighting.h"
+#include "r3d_material.h"
+#include "r3d_mesh_data.h"
+#include "r3d_mesh.h"
+#include "r3d_model.h"
+#include "r3d_particles.h"
+#include "r3d_skeleton.h"
+#include "r3d_skybox.h"
+#include "r3d_sprite.h"
+#include "r3d_utils.h"
+
+#endif // R3D_H
